[Add]GameScene.h.cpp & [Update]GameFramework.cpp

+++ GO_VortexRotation.cpp
@@ -4,8 +4,8 @@
 
 void GO_VortexRotation::Update(void)
 {
-	mp_vortex->SetVortexAngle(atan2(mp_player->GetPos().x - mp_vortex->GetVortexPos().x,
-		mp_player->GetPos().y - mp_vortex->GetVortexPos().y));
+	mp_vortex->SetVortexAngle((double)atan2(mp_player->GetPos().x - (double)mp_vortex->GetVortexPos().x,
+		(double)mp_player->GetPos().y - (double)mp_vortex->GetVortexPos().y));
 }
